app/pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = ["shunk-py <auau3700@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8 <3.9"
django = "^3.0.5"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
django-filter = "^2.2.0"
psycopg2-binary = "^2.8.5"
django-rest-auth = "^0.9.5"
django-allauth = "^0.41.0"
djangorestframework_jwt = "^1.11.0"
django-environ = "^0.4.5"
django-cors-headers = "^3.2.1"
gunicorn = "^20.0.4"
celery = "^4.4.6"
redis = "^3.5.3"
django_celery_beat = "^2.0.0"
django-celery-results = "^1.2.1"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 119
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
